kinesis110: wait to settle after selecting a matrix row
This was uncovered by 4dc04e1b2dbb175f88e3695e84fd55f4d70538e6, which
changed the sequence of row selection vs storage lookup. Previously
the sequence was:

  1. matrix_select_row
  2. storage_read_byte
  3. matrix_read_column

After the change, the sequence became:

  1. matrix_select_row
  2. matrix_read_column
  3. storage_read_byte

The duration of storage_read_byte was sufficient to let the state of
INPUT_PIN settle. Now we add a small delay to have the same effect.